refactor: phase 3 — repository decomposition (architecture seams)
Split monolithic StateManager (1599 lines) into 7 domain repositories
with StateManager as backward-compatible facade (420 lines).

Repositories: sources, fragments, embeddings_store, gaps, citations,
plans, prune — each in src/klemma/repositories/ with shared
BaseRepository(conn_factory) pattern.

Eliminated all private method access from production code:
- cli.py: _conn() → get_existing_source_ids(), get_sources_without_embeddings()
- acquirer.py: _conn() + _set_sections_inline() → set_source_sections()

Cross-repo deps resolved via callable injection (rerank_gaps_semantic).
All 226 tests pass (13 new in test_repositories.py).
